backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pandas==2.1.4
openpyxl==3.1.2
sqlalchemy==2.0.25
pydantic==2.5.3
python-multipart==0.0.6
python-Levenshtein==0.23.0
